bugs fixed

+++ src/main/java/by/itstep/application/repository/TestRepository.java
@@ -10,9 +10,9 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface TestRepository extends JpaRepository<Test, Integer> {
+public interface TestRepository extends JpaRepository<Test, Long> {
     @Query("SELECT DISTINCT t FROM Test t JOIN FETCH t.questions WHERE t.id = :id")
-    Optional<Test> findByIdWithQuestions(@Param("id") Integer id);
+    Optional<Test> findByIdWithQuestions(@Param("id") Long id);
 
     @Query("SELECT DISTINCT t FROM Test t JOIN FETCH t.questions")
     List<Test> findAllWithQuestions();
